i3: Added new logic to brightness control script
The old brightness control script was changing the brightness value
detected by xrandr. The problem with this logic is that it did not allow
you to really set a reasonable level of brightness for the display. If
you tried to cheat it and go beyond 100% the colours looked all odd.

Using this new logic xrandr sets the brightness level for the actual
device, just like GNOME does, and this allows one to reach a decent
level of brightness for the display.